Project as lib: * fix config service * added pack-list script
