+++ package.json
@@ -1,17 +1,18 @@
 {
   "name": "@k-platform/core",
-  "version": "0.0.17",
+  "version": "0.0.23",
   "description": "todo description",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "author": "Alexander Kiriliuk",
   "license": "Apache-2.0",
   "scripts": {
+    "build": "rimraf dist && tsc -p ./src/tsconfig.lib.json",
+    "create:tar": "node ./create-pack-list.js",
+    "publish": "npm publish ./dist/package.tgz",
     "format": "prettier --write \"src/**/*.ts\"",
-    "build": "tsc -p ./src/tsconfig.lib.json",
-    "publish": "npm publish --registry=https://npm.specs-tech.ru/",
     "init": "npm run init:config && npm run init:web-app",
-    "init:config": "nest start init-config",
+    "init:config": "nest start init-config --debug",
     "init:web-app": "nest start init-web-app",
     "build:web-app": "nest build web-app",
     "lint": "eslint \"{src,app,modules,libs,test}/**/*.ts\"",
@@ -96,6 +97,7 @@
     "@nestjs/schematics": "10.1.1",
     "@nestjs/testing": "10.3.7",
     "@nestjs/typeorm": "10.0.2",
+    "@npmcli/arborist": "7.5.1",
     "@types/adm-zip": "0.5.5",
     "@types/express": "4.17.21",
     "@types/jest": "29.5.12",
@@ -108,9 +110,12 @@
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-prettier": "5.1.3",
     "jest": "29.7.0",
+    "npm-packlist": "8.0.2",
     "prettier": "3.2.5",
+    "rimraf": "5.0.5",
     "source-map-support": "0.5.21",
     "supertest": "6.3.4",
+    "tar": "7.0.1",
     "ts-jest": "29.1.2",
     "ts-loader": "9.5.1",
     "ts-node": "10.9.2",
